Move the CSSParser changes
Also deduplicate the CSSParser methods (helped me in development) and
unify a bit between the two chapters
